🎨 (feature/api) Moved S3 bucket to be dependency
